change the typedefs for basic types from Q_(U)INTXX to q(u)intxx, etc. The old types looked ugly and too much like macros and where hard to type all the time.
[git-p4: depot-paths = "//depot/qt/main/": change = 165594]

+++ src/compat/other/q3membuf.cpp
@@ -39,7 +39,7 @@ Q3Membuf::~Q3Membuf()
 */
 bool Q3Membuf::consumeBytes(Q_ULONG nbytes, char *sink)
 {
-    if (nbytes <= 0 || nbytes > _size)
+    if (nbytes <= 0 || (qint64)nbytes > _size)
         return false;
     _size -= nbytes;
     while (!buf.isEmpty()) {
